+++ src/integration-test/java/com/commercetools/sync/integration/externalsource/products/ProductSyncFilterIT.java
@@ -28,6 +28,7 @@ import java.util.Locale;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
+import static com.commercetools.sync.commons.asserts.statistics.AssertionsForStatistics.assertThat;
 import static com.commercetools.sync.integration.commons.utils.CategoryITUtils.OLD_CATEGORY_CUSTOM_TYPE_KEY;
 import static com.commercetools.sync.integration.commons.utils.CategoryITUtils.OLD_CATEGORY_CUSTOM_TYPE_NAME;
 import static com.commercetools.sync.integration.commons.utils.CategoryITUtils.createCategories;
@@ -50,7 +51,6 @@ import static com.commercetools.sync.products.SyncFilter.ofBlackList;
 import static com.commercetools.sync.products.SyncFilter.ofWhiteList;
 import static com.commercetools.tests.utils.CompletionStageUtil.executeBlocking;
 import static io.sphere.sdk.producttypes.ProductType.referenceOfId;
-import static java.lang.String.format;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -142,10 +142,7 @@ public class ProductSyncFilterIT {
         final ProductSyncStatistics syncStatistics =
                 executeBlocking(productSync.sync(singletonList(productDraft)));
 
-
-        assertThat(syncStatistics.getReportMessage())
-                .isEqualTo(format("Summary: %d products were processed in total (%d created, %d updated and %d products"
-                        + " failed to sync).", 1, 0, 1, 0));
+        assertThat(syncStatistics).hasValues(1, 0, 1, 0);
         assertThat(updateActionsFromSync.stream()
                 .noneMatch(updateAction -> updateAction instanceof RemoveFromCategory)).isTrue();
         assertThat(updateActionsFromSync.stream()
@@ -169,9 +166,7 @@ public class ProductSyncFilterIT {
         final ProductSyncStatistics syncStatistics =
                 executeBlocking(productSync.sync(singletonList(productDraft)));
 
-        assertThat(syncStatistics.getReportMessage())
-                .isEqualTo(format("Summary: %d products were processed in total (%d created, %d updated and %d products"
-                        + " failed to sync).", 1, 0, 1, 0));
+        assertThat(syncStatistics).hasValues(1, 0, 1, 0);
         assertThat(updateActionsFromSync).hasSize(1);
         final UpdateAction<Product> updateAction = updateActionsFromSync.get(0);
         assertThat(updateAction.getAction()).isEqualTo("changeName");