+++ src/test/java/com/commercetools/sync/categories/CategorySyncTest.java
@@ -23,6 +23,7 @@ import java.util.concurrent.CompletionException;
 import static com.commercetools.sync.categories.CategorySyncMockUtils.getMockCategoryDraft;
 import static com.commercetools.sync.commons.MockUtils.getMockCategoryService;
 import static com.commercetools.sync.commons.MockUtils.getMockTypeService;
+import static com.commercetools.sync.commons.asserts.statistics.AssertionsForStatistics.assertThat;
 import static java.lang.String.format;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -69,12 +70,7 @@ public class CategorySyncTest {
         final CategorySyncStatistics syncStatistics = mockCategorySync.sync(new ArrayList<>())
                                                                       .toCompletableFuture().join();
 
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(0);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 0 categories were processed in "
-            + "total (0 created, 0 updated, 0 failed to sync and 0 categories with a missing parent).");
+        assertThat(syncStatistics).hasValues( 0, 0, 0, 0);
         assertThat(errorCallBackMessages).isEmpty();
         assertThat(errorCallBackExceptions).isEmpty();
     }
@@ -93,12 +89,8 @@ public class CategorySyncTest {
 
         final CategorySyncStatistics syncStatistics = mockCategorySync.sync(categoryDrafts)
                                                                       .toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 0 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues(1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo("CategoryDraft is null.");
         assertThat(errorCallBackExceptions).hasSize(1);
@@ -120,12 +112,8 @@ public class CategorySyncTest {
 
         final CategorySyncStatistics syncStatistics = mockCategorySync.sync(categoryDrafts)
                                                                       .toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 0 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues(1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo("CategoryDraft with name: "
             + "LocalizedString(en -> noKeyDraft) doesn't have a key.");
@@ -149,12 +137,7 @@ public class CategorySyncTest {
         final CategorySyncStatistics syncStatistics = mockCategorySync.sync(categoryDrafts)
                                                                       .toCompletableFuture().join();
 
-        assertThat(syncStatistics.getCreated()).isEqualTo(1);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (1 created, 0 updated, 0 failed to sync and 1 categories with a missing parent).");
+        assertThat(syncStatistics).hasValues( 1, 1, 0, 0);
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
     }
@@ -167,12 +150,8 @@ public class CategorySyncTest {
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts)
                                                                   .toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(1);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 1 updated, 0 failed to sync and 1 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 1, 0);
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
     }
@@ -195,12 +174,7 @@ public class CategorySyncTest {
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
 
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 0 failed to sync and 1 categories with a missing parent).");
+        assertThat(syncStatistics).hasValues( 1, 0, 0, 0);
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
     }
@@ -216,12 +190,8 @@ public class CategorySyncTest {
             "customTypeId", new HashMap<>()));
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 0 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo(format("Failed to resolve references on CategoryDraft with"
             + " key:'key'. Reason: %s: Failed to resolve parent reference on CategoryDraft"
@@ -242,12 +212,8 @@ public class CategorySyncTest {
             "customTypeId", new HashMap<>()));
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 0 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo(format("Failed to resolve references on CategoryDraft with"
             + " key:'key'. Reason: %s: Failed to resolve parent reference on CategoryDraft"
@@ -271,12 +237,8 @@ public class CategorySyncTest {
         categoryDrafts.add(categoryDraft);
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(1);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 1 updated, 0 failed to sync and 0 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 1, 0);
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
     }
@@ -291,12 +253,8 @@ public class CategorySyncTest {
             "customTypeId", new HashMap<>()));
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 0 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo(format("Failed to resolve references on CategoryDraft with"
             + " key:'key'. Reason: %s: Failed to resolve parent reference on CategoryDraft"
@@ -316,12 +274,8 @@ public class CategorySyncTest {
             "", new HashMap<>()));
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 1 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo(format("Failed to resolve references on CategoryDraft with"
             + " key:'key'. Reason: %s: Failed to resolve custom type reference on "
@@ -343,12 +297,8 @@ public class CategorySyncTest {
             uuidCustomTypeKey, new HashMap<>()));
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(1);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 0 updated, 1 failed to sync and 1 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues( 1, 0, 0, 1);
         assertThat(errorCallBackMessages).hasSize(1);
         assertThat(errorCallBackMessages.get(0)).isEqualTo(format("Failed to resolve references on CategoryDraft with"
             + " key:'key'. Reason: %s: Failed to resolve custom type reference on "
@@ -375,12 +325,8 @@ public class CategorySyncTest {
             uuidCustomTypeKey, new HashMap<>()));
 
         final CategorySyncStatistics syncStatistics = categorySync.sync(categoryDrafts).toCompletableFuture().join();
-        assertThat(syncStatistics.getCreated()).isEqualTo(0);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(1);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(1);
-        assertThat(syncStatistics.getReportMessage()).isEqualTo("Summary: 1 categories were processed in "
-            + "total (0 created, 1 updated, 0 failed to sync and 1 categories with a missing parent).");
+
+        assertThat(syncStatistics).hasValues(1, 0, 1, 0);
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
     }
@@ -457,16 +403,8 @@ public class CategorySyncTest {
         int expectedNumberOfCalls = (int) (Math.ceil(numberOfCategoryDrafts / batchSize) + 1);
         verify(mockCategorySync, times(expectedNumberOfCalls)).syncBatches(any(), any());
 
-        int expectedNumberOfCategoriesCreated = expectedNumberOfCalls - 1;
-        assertThat(syncStatistics.getCreated()).isEqualTo(expectedNumberOfCategoriesCreated);
-        assertThat(syncStatistics.getFailed()).isEqualTo(0);
-        assertThat(syncStatistics.getUpdated()).isEqualTo(0);
-        assertThat(syncStatistics.getProcessed()).isEqualTo(categoryDrafts.size());
-        assertThat(syncStatistics.getReportMessage()).isEqualTo(format("Summary: %s categories were processed"
-                + " in total (%s created, 0 updated, 0 failed to sync and %s categories with a missing parent).",
-            categoryDrafts.size(),
-            expectedNumberOfCategoriesCreated,
-            categoryDrafts.size()));
+        assertThat(syncStatistics)
+            .hasValues( categoryDrafts.size(), expectedNumberOfCalls - 1, 0, 0, categoryDrafts.size());
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
 
@@ -493,17 +431,11 @@ public class CategorySyncTest {
         verify(mockCategorySyncWithDefaultBatchSize, times(expectedNumberOfCalls)).syncBatches(any(), any());
 
 
-        expectedNumberOfCategoriesCreated = expectedNumberOfCalls - 1;
+        final int expectedNumberOfCategoriesCreated = expectedNumberOfCalls - 1;
         final int expectedFailedCategories = categoryDrafts.size() - (expectedNumberOfCategoriesCreated);
+        assertThat(syncStatisticsWithDefaultBatchSize).hasValues(categoryDrafts.size(),
+            expectedNumberOfCategoriesCreated, 0, expectedFailedCategories, categoryDrafts.size());
 
-        assertThat(syncStatisticsWithDefaultBatchSize.getCreated()).isEqualTo(expectedNumberOfCategoriesCreated);
-        assertThat(syncStatisticsWithDefaultBatchSize.getFailed()).isEqualTo(expectedFailedCategories);
-        assertThat(syncStatisticsWithDefaultBatchSize.getUpdated()).isEqualTo(0);
-        assertThat(syncStatisticsWithDefaultBatchSize.getProcessed()).isEqualTo(categoryDrafts.size());
-        assertThat(syncStatisticsWithDefaultBatchSize.getReportMessage())
-            .isEqualTo(format("Summary: %s categories were processed in total (%s created, 0 updated, %s failed to sync"
-                    + " and %s categories with a missing parent).", categoryDrafts.size(),
-                expectedNumberOfCategoriesCreated, expectedFailedCategories, categoryDrafts.size()));
         assertThat(errorCallBackMessages).hasSize(0);
         assertThat(errorCallBackExceptions).hasSize(0);
     }